+++ src/symfluence/models/fuse/forcing_adapter.py
@@ -13,8 +13,8 @@ from typing import Callable, Dict, List
 
 import xarray as xr
 
+from symfluence.core.modeling.adapters import ForcingAdapter
 from symfluence.core.registries import R
-from symfluence.models.adapters import ForcingAdapter
 
 
 @R.forcing_adapters.add('FUSE')